first compiling full conversion

+++ src/operations.h
@@ -19,7 +19,8 @@ struct parsed_contract {
 #define PROTOCOL_HASH_SIZE 32
 
 struct parsed_proposal {
-    int32_t voting_period;
+    uint64_t voting_period; // must be 64 bit for function call reasons
+    // TODO: Make 32 bit version of number_to_string_indirect
     uint8_t protocol_hash[PROTOCOL_HASH_SIZE];
 };
